[#14] Refactor the HTTP client
